+++ src/app/modules/noi-dung/phong-gia/date-book-room/date-book-room.component.html
@@ -1,0 +1,124 @@
+<div class="table-scroll">
+    <table class="table table-bordered my-tbody">
+        <tr>
+            <td><label>Thứ 2</label><label>12/10</label></td>
+            <td><label>Thứ 2</label><label>12/10</label></td>
+            <td><label>Thứ 2</label><label>12/10</label></td>
+            <td><label>Thứ 2</label><label>12/10</label></td>
+            <td><label>Thứ 2</label><label>12/10</label></td>
+            <td><label>Thứ 2</label><label>12/10</label></td>
+            <td><label>Thứ 2</label><label>12/10</label></td>
+            <td><label>Thứ 2</label><label>12/10</label></td>
+            <td><label>Thứ 2</label><label>12/10</label></td>
+            <td><label>Thứ 2</label><label>12/10</label></td>
+            <td><label>Thứ 2</label><label>12/10</label></td>
+            <td><label>Thứ 2</label><label>12/10</label></td>
+            <td><label>Thứ 2</label><label>12/10</label></td>
+            <td><label>Thứ 2</label><label>12/10</label></td>
+            <td><label>Thứ 2</label><label>12/10</label></td>
+        </tr>
+        <tr>
+            <td> <mat-slide-toggle></mat-slide-toggle></td>
+            <td> <mat-slide-toggle></mat-slide-toggle></td>
+            <td> <mat-slide-toggle></mat-slide-toggle></td>
+            <td> <mat-slide-toggle></mat-slide-toggle></td>
+            <td> <mat-slide-toggle></mat-slide-toggle></td>
+            <td> <mat-slide-toggle></mat-slide-toggle></td>
+            <td> <mat-slide-toggle></mat-slide-toggle></td>
+            <td> <mat-slide-toggle></mat-slide-toggle></td>
+            <td> <mat-slide-toggle></mat-slide-toggle></td>
+            <td> <mat-slide-toggle></mat-slide-toggle></td>
+            <td> <mat-slide-toggle></mat-slide-toggle></td>
+            <td> <mat-slide-toggle></mat-slide-toggle></td>
+            <td> <mat-slide-toggle></mat-slide-toggle></td>
+            <td> <mat-slide-toggle></mat-slide-toggle></td>
+            <td> <mat-slide-toggle></mat-slide-toggle></td>
+        </tr>
+        <tr>
+            <td><label class="label-price">10</label></td>
+            <td><label class="label-price">10</label></td>
+            <td><label class="label-price">10</label></td>
+            <td><label class="label-price">10</label></td>
+            <td><label class="label-price">10</label></td>
+            <td><label class="label-price">10</label></td>
+            <td><label class="label-price">10</label></td>
+            <td><label class="label-price">10</label></td>
+            <td><label class="label-price">10</label></td>
+            <td><label class="label-price">10</label></td>
+            <td><label class="label-price">10</label></td>
+            <td><label class="label-price">10</label></td>
+            <td><label class="label-price">10</label></td>
+            <td><label class="label-price">10</label></td>
+            <td><label class="label-price">10</label></td>
+        </tr>
+        <tr>
+            <td><label class="label-price">10</label></td>
+            <td><label class="label-price">10</label></td>
+            <td><label class="label-price">10</label></td>
+            <td><label class="label-price">10</label></td>
+            <td><label class="label-price">10</label></td>
+            <td><label class="label-price">10</label></td>
+            <td><label class="label-price">10</label></td>
+            <td><label class="label-price">10</label></td>
+            <td><label class="label-price">10</label></td>
+            <td><label class="label-price">10</label></td>
+            <td><label class="label-price">10</label></td>
+            <td><label class="label-price">10</label></td>
+            <td><label class="label-price">10</label></td>
+            <td><label class="label-price">10</label></td>
+            <td><label class="label-price">10</label></td>
+        </tr>
+        <tr>
+            <td><label>4</label></td>
+            <td><label>3</label></td>
+            <td><label>2</label></td>
+            <td><label>10</label></td>
+            <td><label>10</label></td>
+            <td><label>10</label></td>
+            <td><label>10</label></td>
+            <td><label>10</label></td>
+            <td><label>10</label></td>
+            <td><label>10</label></td>
+            <td><label>10</label></td>
+            <td><label>10</label></td>
+            <td><label>10</label></td>
+            <td><label>10</label></td>
+            <td><label>10</label></td>
+        </tr>
+        <tr>
+            <td><label>4</label></td>
+            <td><label>3</label></td>
+            <td><label>2</label></td>
+            <td><label>10</label></td>
+            <td><label>10</label></td>
+            <td><label>10</label></td>
+            <td><label>10</label></td>
+            <td><label>10</label></td>
+            <td><label>10</label></td>
+            <td><label>10</label></td>
+            <td><label>10</label></td>
+            <td><label>10</label></td>
+            <td><label>10</label></td>
+            <td><label>10</label></td>
+            <td><label>10</label></td>
+        </tr>
+        <tr>
+            <td><label>4</label></td>
+            <td><label>3</label></td>
+            <td><label>2</label></td>
+            <td><label>10</label></td>
+            <td><label>10</label></td>
+            <td><label>10</label></td>
+            <td><label>10</label></td>
+            <td><label>10</label></td>
+            <td><label>10</label></td>
+            <td><label>10</label></td>
+            <td><label>10</label></td>
+            <td><label>10</label></td>
+            <td><label>10</label></td>
+            <td><label>10</label></td>
+            <td><label>10</label></td>
+        </tr>
+    
+    </table>
+</div>
