+++ src/app/modules/noi-dung/phong-gia/date-book-room/date-book-room.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-date-book-room',
+  templateUrl: './date-book-room.component.html',
+  styleUrls: ['./date-book-room.component.scss']
+})
+export class DateBookRoomComponent implements OnInit {
+  isChecked = true;
+  constructor() { }
+
+  ngOnInit(): void {
+  }
+
+}
