+++ src/app/modules/noi-dung/phong-gia/list-date-book-room/list-date-book-room.component.scss
@@ -1,0 +1,49 @@
+.label-icon{
+    label {
+        border: 1px solid #dddddd;
+        border-radius:20px ;
+        padding: 2px 10px;
+        margin-right: 15px;
+        margin-bottom: 15px;
+        span.green{
+            color: green;
+            padding-right: 10px;
+            line-height: 28px;
+        }
+        span.red{
+            color: #8a8e91;
+            padding-right: 10px;
+            line-height: 28px;
+        }
+        span.yellow{
+            color: yellow;
+            padding-right: 10px;
+            line-height: 28px;
+        }
+        span.text{
+            position: relative;
+            top: -4px;
+        }
+    }
+}
+
+.my-th-table {
+    table-layout: fixed;
+
+    tr{
+        width: 220.8px;
+        td{
+            border: 1px solid #dddddd;
+        }
+    }
+    tr:nth-child(1){
+        height: 71.4px;
+       
+    }
+
+}
+
+.btn-right
+    button{
+        margin-right: 15px;
+    }
