+++ src/app/modules/noi-dung/phong-gia/phong-gia.component.scss
@@ -13,82 +13,7 @@
     flex: 0 0 80%;
     max-width: 80%;
 }
-.label-icon{
-    label {
-        border: 1px solid #dddddd;
-        border-radius:20px ;
-        padding: 2px 10px;
-        margin-right: 15px;
-        margin-bottom: 15px;
-        span.green{
-            color: green;
-            padding-right: 10px;
-            line-height: 28px;
-        }
-        span.red{
-            color: #8a8e91;
-            padding-right: 10px;
-            line-height: 28px;
-        }
-        span.yellow{
-            color: yellow;
-            padding-right: 10px;
-            line-height: 28px;
-        }
-        span.text{
-            position: relative;
-            top: -4px;
-        }
-    }
-}
-.table-scroll {
-    overflow-x: auto;
-    width: 100%;
 
-    .my-tbody {
-        table-layout: fixed;
-
-        td {
-            width: 80px;
-            border: 1px solid #dddddd;
-            text-align: center; 
-            vertical-align: middle;
-            .label-price{
-                background: #eee;
-                padding: 8px 25px;
-                margin: 0;
-            }
-        }
-
-        tr:nth-child(3), tr:nth-child(4){
-            height: 56px;
-        }
-        tr:nth-child(5), tr:nth-child(6), tr:nth-child(7){
-            height: 58.4px;
-        }
-    }
-}
-
-.my-th-table {
-    table-layout: fixed;
-
-    tr{
-        width: 220.8px;
-        td{
-            border: 1px solid #dddddd;
-        }
-    }
-    tr:nth-child(1){
-        height: 71.4px;
-       
-    }
-
-}
-
-.btn-right
-    button{
-        margin-right: 15px;
-    }
 .btn-search {
     margin-left: 0;
     mat-form-field {
@@ -101,18 +26,3 @@
 }
 
 
-.table-scroll::-webkit-scrollbar {
-    width: 1em;
-    height: 10px;
-}
-
-.table-scroll::-webkit-scrollbar-track {
-    box-shadow: inset 0 0 6px rgba(0, 0, 0, 0.3);
-    border-radius: 20px;
-}
-
-.table-scroll::-webkit-scrollbar-thumb {
-    background-color: #3699ff;
-    outline: 1px solid #3699ff;
-    border-radius: 20px;
-}
